requirements.txt
python-telegram-bot==20.7
python-pptx==0.6.23
python-dotenv==1.0.0
gigachat==0.1.28
pdfplumber==0.11.0
Pillow==10.3.0
